Use the css from simplecss.org and switch to HTML 5
By default, the Org publisher doesn't use HTML 5, so this commit
switches to HTML 5, uses a preconfigured stylesheet from
simplecss.org,
and changes the preamble, content, and postamble CARS so that
simplecss works with correct styling. It doesn't like the default divs
that are inserted by preamble, content, and postamble.